+++ packages/utils/src/handleInterceptorCode/index.ts
@@ -1,4 +1,7 @@
 import HandleInterceptorCode from './handleInterceptorCode';
+import HandleInterceptorCodeCom from './HandleInterceptorCodeCom';
 export type { IHandleCodeParams } from './handleInterceptorCode';
+
+export { HandleInterceptorCodeCom };
 
 export default HandleInterceptorCode;